Внести корректировки в классы-контроллеры, доработать классы для хранения данных и бизнес-логики
