+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.util.HashSet;
@@ -9,22 +10,12 @@ import java.util.Set;
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class User {
     private Long id;
     private String email;
     private String login;
     private String name;
     private LocalDate birthday;
-    private Set<Long> friends;
-
-    public Set<Long> getFriends() {
-        if (friends == null) {
-            return friends = new HashSet<>();
-        }
-        return friends;
-    }
-
-    public void setFriends(Set<Long> friends) {
-        this.friends = friends != null ? friends : new HashSet<>();
-    }
+    private Set<Long> friends = new HashSet<>();
 }
