+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -10,6 +10,7 @@ import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
+import java.time.LocalDate;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.stream.Collectors;
@@ -19,6 +20,8 @@ import java.util.stream.Collectors;
 public class FilmService {
     private final FilmStorage filmStorage;
     private final UserStorage userStorage;
+    private static final int MAX_DESCRIPTION_LENGTH = 200;
+    private static final LocalDate DECEMBER_1895 = LocalDate.of(1895,12,28);
 
     @Autowired
     public FilmService(FilmStorage filmStorage, UserStorage userStorage) {
@@ -26,43 +29,45 @@ public class FilmService {
         this.userStorage = userStorage;
     }
 
+    public Film createFilm(Film film) {
+        validateFilm(film);
+        return filmStorage.createFilm(film);
+    }
+
+    public Film updateFilm(Film newFilm) {
+        validateFilm(newFilm);
+        return filmStorage.updateFilm(newFilm);
+    }
+
+    public Collection<Film> getAllFilms() {
+        return filmStorage.getAllFilms();
+    }
+
+    public Film getFilm(Long id) {
+        return filmStorage.getFilm(id);
+    }
+
     public void addLike(Long id, Long userId) {
         Film film = filmStorage.getFilm(id);
-        User user = userStorage.getUser(userId);
-        if (film != null && user != null) {
-            if (film.getLikes().contains(userId)) {
-                log.warn("Попытка повторно поставить лайк фильму");
-                throw new ValidationException("Фильм можно лайкнуть только один раз!");
-            }
-
-            film.getLikes().add(userId);
-        } else if (film == null) {
-            log.warn("Фильм с id = {} не найден", id);
-            throw new NotFoundException("Фильм с id = " + id + " не найден");
-        } else {
-            log.warn("Пользователь с userId = {} не найден", userId);
-            throw new NotFoundException("Пользователь с userId = " + userId + " не найден");
+        validateLike(id, userId);
+        if (film.getLikes().contains(userId)) {
+            log.warn("Попытка повторно поставить лайк фильму");
+            throw new ValidationException("Фильм можно лайкнуть только один раз!");
         }
+
+        film.getLikes().add(userId);
     }
 
     public void deleteLike(Long id, Long userId) {
         Film film = filmStorage.getFilm(id);
-        User user = userStorage.getUser(userId);
-        if (film != null && user != null) {
-            if (!film.getLikes().contains(userId)) {
-                log.warn("Попытка удалить несуществующий лайк. Ранее лайк этому фильму не был поставлен");
-                throw new ValidationException("Нельзя удалить несуществующий лайк. " +
-                        "Ранее лайк этому фильму не был поставлен");
+        validateLike(id, userId);
+        if (!film.getLikes().contains(userId)) {
+            log.warn("Попытка удалить несуществующий лайк. Ранее лайк этому фильму не был поставлен");
+            throw new ValidationException("Нельзя удалить несуществующий лайк. " +
+                    "Ранее лайк этому фильму не был поставлен");
             }
 
-            film.getLikes().remove(userId);
-        } else if (film == null) {
-            log.warn("Фильм с id = {} не найден", id);
-            throw new NotFoundException("Фильм с id = " + id + " не найден");
-        } else {
-            log.warn("Пользователь с userId = {} не найден", userId);
-            throw new NotFoundException("Пользователь с userId = " + userId + " не найден");
-        }
+        film.getLikes().remove(userId);
     }
 
     public Collection<Film> getPopularFilms(int count) {
@@ -71,5 +76,40 @@ public class FilmService {
                 .sorted(Comparator.comparingInt((Film film) -> film.getLikes().size()).reversed())
                 .limit(count)
                 .collect(Collectors.toList());
+    }
+
+    public void validateFilm(Film film) {
+        log.debug("Начало проверки соответствия данных фильма {} всем критериям.", film.getName());
+
+        if (film.getName() == null || film.getName().isBlank()) {
+            log.warn("Название фильма {} не соответствует требованиям.", film.getName());
+            throw new ValidationException("Название не может быть пустым.");
+        }
+        if (film.getDescription() == null || film.getDescription().length() > MAX_DESCRIPTION_LENGTH) {
+            log.warn("Описание фильма {} не соответствует требованиям.", film.getDescription());
+            throw new ValidationException("Максимальная длина описания — 200 символов.");
+        }
+        if (film.getReleaseDate() == null || film.getReleaseDate().isBefore(DECEMBER_1895)) {
+            log.warn("Дата выпуска фильма {} не соответствует требованиям.", film.getReleaseDate());
+            throw new ValidationException("Дата релиза должна быть не раньше 28 декабря 1895 года.");
+        }
+        if (film.getDuration() == null || !film.getDuration().isPositive()) {
+            log.warn("Продолжительность фильма {} не соответствует требованиям.", film.getDuration());
+            throw new ValidationException("Продолжительность фильма должна быть положительным числом.");
+        }
+
+        log.debug("Проверка данных фильма {} прошла успешно.", film.getName());
+    }
+
+    public void validateLike(Long id, Long userId) {
+        Film film = filmStorage.getFilm(id);
+        User user = userStorage.getUser(userId);
+        if (film == null) {
+            log.warn("Фильм с id = {} не найден", id);
+            throw new NotFoundException("Фильм с id = " + id + " не найден");
+        } else if (user == null) {
+            log.warn("Пользователь с userId = {} не найден", userId);
+            throw new NotFoundException("Пользователь с userId = " + userId + " не найден");
+        }
     }
 }
