+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -3,10 +3,8 @@ package ru.yandex.practicum.filmorate.storage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 
-import java.time.LocalDate;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
@@ -16,12 +14,9 @@ import java.util.Map;
 public class InMemoryFilmStorage implements FilmStorage {
     private final Map<Long, Film> films = new HashMap<>();
     private static long id = 0;
-    private static final int MAX_DESCRIPTION_LENGTH = 200;
-    private static final LocalDate DEСEMBER_1895 = LocalDate.of(1895,12,28);
 
     @Override
     public Film createFilm(Film film) {
-        validateFilm(film);
         film.setId(++id);
         films.put(film.getId(), film);
         log.debug("Фильм {} успешно добавлен", film.getName());
@@ -31,7 +26,6 @@ public class InMemoryFilmStorage implements FilmStorage {
     @Override
     public Film updateFilm(Film newFilm) {
         if (films.containsKey(newFilm.getId())) {
-            validateFilm(newFilm);
             films.put(newFilm.getId(), newFilm);
             log.debug("Фильм с id {} успешно обновлен.", newFilm.getId());
             return newFilm;
@@ -56,27 +50,5 @@ public class InMemoryFilmStorage implements FilmStorage {
         }
     }
 
-    @Override
-    public void validateFilm(Film film) {
-        log.debug("Начало проверки соответствия данных фильма {} всем критериям.", film.getName());
 
-        if (film.getName() == null || film.getName().isBlank()) {
-            log.warn("Название фильма {} не соответствует требованиям.", film.getName());
-            throw new ValidationException("Название не может быть пустым.");
-        }
-        if (film.getDescription() == null || film.getDescription().length() > MAX_DESCRIPTION_LENGTH) {
-            log.warn("Описание фильма {} не соответствует требованиям.", film.getDescription());
-            throw new ValidationException("Максимальная длина описания — 200 символов.");
-        }
-        if (film.getReleaseDate() == null || film.getReleaseDate().isBefore(DEСEMBER_1895)) {
-            log.warn("Дата выпуска фильма {} не соответствует требованиям.", film.getReleaseDate());
-            throw new ValidationException("Дата релиза должна быть не раньше 28 декабря 1895 года.");
-        }
-        if (film.getDuration() == null || !film.getDuration().isPositive()) {
-            log.warn("Продолжительность фильма {} не соответствует требованиям.", film.getDuration());
-            throw new ValidationException("Продолжительность фильма должна быть положительным числом.");
-        }
-
-        log.debug("Проверка данных фильма {} прошла успешно.", film.getName());
-    }
 }
